LAB-1398: Delete wait shims and dead exports
