+++ internal/mux/pane_meta_kv.go
@@ -201,10 +201,3 @@ func RemovePaneMetaKV(meta *PaneMeta, key string) (manualBranch bool, err error)
 	*meta = next
 	return manualBranch, nil
 }
-func FormatTrackedPRsValue(prs []proto.TrackedPR) string {
-	return formatTrackedPRsValue(prs)
-}
-
-func FormatTrackedIssuesValue(issues []proto.TrackedIssue) string {
-	return formatTrackedIssuesValue(issues)
-}
