+++ internal/server/commands_wait_test.go
@@ -8,160 +8,6 @@ import (
 	"github.com/weill-labs/amux/internal/proto"
 )
 
-func TestParseWaitArgs(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name         string
-		args         []string
-		wantAfter    uint64
-		wantAfterSet bool
-		wantTimeout  time.Duration
-		wantErr      string
-	}{
-		{
-			name:         "defaults",
-			wantTimeout:  3 * time.Second,
-			wantAfterSet: false,
-		},
-		{
-			name:         "after and timeout",
-			args:         []string{"--after", "7", "--timeout", "25ms"},
-			wantAfter:    7,
-			wantAfterSet: true,
-			wantTimeout:  25 * time.Millisecond,
-		},
-		{
-			name:    "missing after value",
-			args:    []string{"--after"},
-			wantErr: "missing value for --after",
-		},
-		{
-			name:    "invalid after value",
-			args:    []string{"--after", "bogus"},
-			wantErr: "invalid value for --after: bogus",
-		},
-		{
-			name:    "missing timeout value",
-			args:    []string{"--timeout"},
-			wantErr: "missing value for --timeout",
-		},
-		{
-			name:    "invalid timeout value",
-			args:    []string{"--timeout", "later"},
-			wantErr: "invalid value for --timeout: later",
-		},
-		{
-			name:    "unknown flag",
-			args:    []string{"--wat"},
-			wantErr: "unknown flag: --wat",
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			afterGen, afterSet, timeout, err := parseWaitArgs(tt.args)
-			if tt.wantErr != "" {
-				if err == nil || err.Error() != tt.wantErr {
-					t.Fatalf("parseWaitArgs(%v) error = %v, want %q", tt.args, err, tt.wantErr)
-				}
-				return
-			}
-			if err != nil {
-				t.Fatalf("parseWaitArgs(%v): %v", tt.args, err)
-			}
-			if afterGen != tt.wantAfter || afterSet != tt.wantAfterSet || timeout != tt.wantTimeout {
-				t.Fatalf("parsed = (%d, %t, %v), want (%d, %t, %v)", afterGen, afterSet, timeout, tt.wantAfter, tt.wantAfterSet, tt.wantTimeout)
-			}
-		})
-	}
-}
-
-func TestParseWaitArgsWithDefault(t *testing.T) {
-	t.Parallel()
-
-	afterGen, afterSet, timeout, err := parseWaitArgsWithDefault(nil, 15*time.Second)
-	if err != nil {
-		t.Fatalf("parseWaitArgsWithDefault(nil): %v", err)
-	}
-	if afterGen != 0 || afterSet || timeout != 15*time.Second {
-		t.Fatalf("parsed = (%d, %t, %v), want (0, false, 15s)", afterGen, afterSet, timeout)
-	}
-}
-
-func TestParseTimeout(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name        string
-		args        []string
-		startIdx    int
-		defaultTime time.Duration
-		want        time.Duration
-		wantErr     string
-	}{
-		{
-			name:        "default timeout",
-			args:        []string{"pane-1"},
-			startIdx:    1,
-			defaultTime: 5 * time.Second,
-			want:        5 * time.Second,
-		},
-		{
-			name:        "explicit timeout",
-			args:        []string{"pane-1", "--timeout", "25ms"},
-			startIdx:    1,
-			defaultTime: 5 * time.Second,
-			want:        25 * time.Millisecond,
-		},
-		{
-			name:        "invalid timeout",
-			args:        []string{"pane-1", "--timeout", "later"},
-			startIdx:    1,
-			defaultTime: 5 * time.Second,
-			wantErr:     "invalid value for --timeout: later",
-		},
-		{
-			name:        "missing timeout value",
-			args:        []string{"pane-1", "--timeout"},
-			startIdx:    1,
-			defaultTime: 5 * time.Second,
-			wantErr:     "missing value for --timeout",
-		},
-		{
-			name:        "unknown flag",
-			args:        []string{"pane-1", "--bogus"},
-			startIdx:    1,
-			defaultTime: 5 * time.Second,
-			wantErr:     "unknown flag: --bogus",
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			got, err := parseTimeout(tt.args, tt.startIdx, tt.defaultTime)
-			if tt.wantErr != "" {
-				if err == nil || err.Error() != tt.wantErr {
-					t.Fatalf("parseTimeout(%v) error = %v, want %q", tt.args, err, tt.wantErr)
-				}
-				return
-			}
-			if err != nil {
-				t.Fatalf("parseTimeout(%v): %v", tt.args, err)
-			}
-			if got != tt.want {
-				t.Fatalf("timeout = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
 func TestHasAfterFlag(t *testing.T) {
 	t.Parallel()
