+++ internal/server/wait_test.go
@@ -348,91 +348,3 @@ func TestBeginPaneOutputWait(t *testing.T) {
 		}
 	})
 }
-
-func TestWaitBusyForegroundProcessGroup(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name   string
-		status mux.ForegroundJobState
-		want   int
-	}{
-		{
-			name:   "idle",
-			status: mux.ForegroundJobState{Idle: true, ForegroundProcessGroup: 12},
-			want:   0,
-		},
-		{
-			name:   "no foreground process group",
-			status: mux.ForegroundJobState{Idle: false},
-			want:   0,
-		},
-		{
-			name:   "foreground process group is returned",
-			status: mux.ForegroundJobState{ForegroundProcessGroup: 56},
-			want:   56,
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			if got := waitBusyForegroundProcessGroup(tt.status); got != tt.want {
-				t.Fatalf("waitBusyForegroundProcessGroup(%+v) = %d, want %d", tt.status, got, tt.want)
-			}
-		})
-	}
-}
-
-func TestWaitBusyReady(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name         string
-		candidatePID int
-		status       mux.ForegroundJobState
-		wantNext     int
-		wantReady    bool
-	}{
-		{
-			name:         "zero candidate does not satisfy readiness",
-			candidatePID: 0,
-			status:       mux.ForegroundJobState{ForegroundProcessGroup: 91},
-			wantNext:     91,
-			wantReady:    false,
-		},
-		{
-			name:         "different foreground process group updates candidate",
-			candidatePID: 91,
-			status:       mux.ForegroundJobState{ForegroundProcessGroup: 104},
-			wantNext:     104,
-			wantReady:    false,
-		},
-		{
-			name:         "same foreground process group is ready",
-			candidatePID: 104,
-			status:       mux.ForegroundJobState{ForegroundProcessGroup: 104},
-			wantNext:     104,
-			wantReady:    true,
-		},
-		{
-			name:         "idle clears candidate",
-			candidatePID: 104,
-			status:       mux.ForegroundJobState{Idle: true, ForegroundProcessGroup: 104},
-			wantNext:     0,
-			wantReady:    false,
-		},
-	}
-
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			gotNext, gotReady := waitBusyReady(tt.candidatePID, tt.status)
-			if gotNext != tt.wantNext || gotReady != tt.wantReady {
-				t.Fatalf("waitBusyReady(%d, %+v) = (%d, %t), want (%d, %t)", tt.candidatePID, tt.status, gotNext, gotReady, tt.wantNext, tt.wantReady)
-			}
-		})
-	}
-}
